Add files via upload
This is a simple tool call explorer meant for Qwen 30b a3b Instruct: https://huggingface.co/Qwen/Qwen3-Coder-30B-A3B-Instruct

Open this simple HTML and you can adjust settings to connect to an AI server that offers Qwen (local VLLM tested). You can add an optional API key if needed. To use, click on a quick prompt for VLLM (like calculator, edit text, etc), and call VLLM streaming or non-streaming. If you are not connected to an AI, you can click the Simulate Streaming button above to simulate a call to an AI, which will respond with an actual qwen 3 30b a3b response to that tool call to explore.

The Build New Tool feature has a small agentic loop that will attempt to build a tool based on your input. Give it details about an api (say, for example, an image generator API), and ask for a tool to call it to generate an image, and it will create and validate a tool call for that purpose to test.